+++ vscode-extension/src/lib/functions.ts
@@ -1,6 +1,8 @@
 import path = require("path");
 import { commands, Uri } from "vscode";
+import { getCachedManifestFromAppId } from "./AppManifest";
 import { ALRange } from "./types";
+import { UI } from "./UI";
 
 export function showDocument(document: string) {
     commands.executeCommand("markdown.showPreview", Uri.file(path.join(__dirname, `../../docs/${document}.md`)));
@@ -12,4 +14,19 @@ export function getRangeForId<T extends ALRange>(id: number, ranges: T[]): T | u
             return range;
         }
     }
+}
+
+export function getPoolIdFromAppIdIfAvailable(appId: string): string {
+    const manifest = getCachedManifestFromAppId(appId);
+    const { appPoolId } = manifest.ninja.config;
+    if (!appPoolId) {
+        return appId;
+    }
+
+    if (appPoolId.length !== 64 || !/[0-9A-Fa-f]{6}/g.test(appPoolId)) {
+        UI.pool.showInvalidAppPoolIdError(manifest);
+        return appId;
+    }
+
+    return appPoolId;
 }